move __swapEndianInt() from data_reader_mnist.cpp to file_utils.hpp to expose and reuse the function move load_file() from tools/compute_mean to file_utils.hpp for reusing the function
