+++ src/data_readers/data_reader_mnist.cpp
@@ -27,11 +27,8 @@
 ////////////////////////////////////////////////////////////////////////////////
 
 #include "lbann/data_readers/data_reader_mnist.hpp"
+#include "lbann/utils/file_utils.hpp"
 #include <cstdio>
-
-inline void __swapEndianInt(unsigned int& ui) {
-  ui = ((ui >> 24) | ((ui<<8) & 0x00FF0000) | ((ui>>8) & 0x0000FF00) | (ui << 24));
-}
 
 namespace lbann {
